Bugfixes and new in-memory providers
Bugfixes for when a word is smaller than the specified n-gram, and
another bugfix for when the latest n-gram wasnt taken into account.
New super-generic in-memory providers make usage easier than ever by
simply passing a text-string.
Interface updated to allow passing an object-parameter.

Acme.MySiteItems has been renamed/upgraded to
Karagulmez.Text.Fuzzy.DefaultProviders to indicate that it can be used
as-is, custom code for a provider isn't necessary (it never really was,
really). This contains all the current providers, including the
in-memory providers.